corrigindo alguns bugs

+++ app/compras/model.py
@@ -5,7 +5,7 @@ from ..users.model import User
 class Compra(db.Model):
     __tablename__ = 'compra'
     id = db.Column(db.Integer,primary_key=True)
-    preço = db.Column(db.Integer)
+    preço = db.Column(db.Float)
     
     produtos = db.relationship('Produto')
     user_id = db.Column(db.Integer,db.ForeignKey('user.id'))
